refactor(app): changed default landing screen
